Move BattleStateObserver
